requirements.txt
python-dotenv
psycopg2-binary
neo4j
tree-sitter-language-pack
watchdog
transformers
torch
torchvision
torchaudio
gitpython
aiofiles
redis